+++ resources/migrations/20211004130940-problems.up.sql
@@ -1,8 +1,8 @@
 create table problems (
   id serial primary key,
-  num integer, -- 本来 unique だが、順番替えしたくなる時もある。
+  num integer,
   is_avail BOOLEAN default true,
   problem text not null,
-  test text default "",
+  test text default '',
   create_at timestamp default current_timestamp,
   update_at timestamp default current_timestamp);